+++ packages/interactions/route-active/types/index.d.ts
@@ -1,6 +1,6 @@
 import { Interaction, Params } from "@curi/router";
-import { HickoryLocation } from "@hickory/root";
-export declare type LocationCheck = (l: HickoryLocation) => boolean;
+import { SessionLocation } from "@hickory/root";
+export declare type LocationCheck = (l: SessionLocation) => boolean;
 export interface ActiveCheckOptions {
     params?: Params;
     partial?: boolean;