+++ packages/react-universal/types/hooks/useBlock.d.ts
@@ -1,2 +1,0 @@
-import { ConfirmationFunction } from "@hickory/root";
-export default function useBlock(active: boolean, fn: ConfirmationFunction): void;