+++ packages/vue/src/index.ts
@@ -1,4 +1,3 @@
-export { BlockComponent } from "./Block";
 export { LinkComponent } from "./Link";
 export { ReactiveResponse } from "./interface";
 export { CuriPluginOptions } from "./plugin";